pip-requirements.txt
flask
python-dotenv # For supporting .flaskenv environment variables
pyyaml
markdown2
pygments
python-frontmatter
